Sizing Update
Layouts for Normal are done and Large seems to work as well. Did not
edit small as its tiny and I pretty much can't edit it. Did not mess
with XLarge either as no test subject is available. However, it seems
that Android will pull the layout files from the next smallest size
folder to use in absence of its own. So XLarge should pull from Large
and Large will pull from Normal. (NOTE) there is no layout-small because
it would seem that Android auto assigned the layout folder to small in
he absence of one.